changes in keeping track of token balances

+++ Advisor-Builder-JAVA/sampleProject/src/main/java/io/yieldster/sampleProject/model/MoveStep.java
@@ -4,6 +4,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.Value;
 
+import java.math.BigInteger;
+
 @Builder
 @Value
 @Data
@@ -11,5 +13,5 @@ public class MoveStep extends Step {
     String type = "MoveStep";
     String fromAsset;
     String toAsset;
-    Double amount;
+    Long amount;
 }
